# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "シーシャSNSアプリケーションのバックエンドAPI\nこのAPIはシーシャの投稿、ユーザー管理を行います",
        "title": "Go-Shisha API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/illionillion/go-shisha"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "メールアドレスとパスワードでログインし、JWT（Cookie）を発行する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "ログイン",
                "parameters": [
                    {
                        "description": "ログイン情報",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ログイン成功",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "バリデーションエラー",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "認証失敗",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cookieを削除し、Refresh Tokenを無効化する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "ログアウト",
                "responses": {
                    "200": {
                        "description": "ログアウト成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "認証失敗",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "認証されたユーザーの情報を取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "現在のユーザー情報取得",
                "responses": {
                    "200": {
                        "description": "ユーザー情報",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.AuthResponse"
                        }
                    },
                    "401": {
                        "description": "認証失敗",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Refresh Tokenを使って新しいAccess Tokenを発行する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "アクセストークンのリフレッシュ",
                "responses": {
                    "200": {
                        "description": "リフレッシュ成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "認証失敗",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "新しいユーザーを登録する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "ユーザー登録",
                "parameters": [
                    {
                        "description": "ユーザー登録情報",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.CreateUserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "登録成功",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "バリデーションエラー",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "メールアドレス重複",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/flavors": {
            "get": {
                "description": "全てのフレーバーの一覧を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flavors"
                ],
                "summary": "フレーバー一覧取得",
                "responses": {
                    "200": {
                        "description": "フレーバー一覧",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/go-shisha-backend_internal_models.Flavor"
                            }
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "description": "全ての投稿の一覧を取得します（総数付き）。認証済みの場合、各投稿のいいね状態（is_liked）を含みます",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "投稿一覧取得",
                "responses": {
                    "200": {
                        "description": "投稿一覧と総数",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.PostsResponse"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "新しい投稿を作成します（認証必須）。注意: slides内のflavor_idが無効な場合、そのスライドはFlavorなしで作成されます（エラーにはなりません）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "投稿作成",
                "parameters": [
                    {
                        "description": "投稿情報",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.CreatePostInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "作成された投稿",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.Post"
                        }
                    },
                    "400": {
                        "description": "バリデーションエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "description": "指定されたIDの投稿情報を取得します。認証済みの場合、いいね状態（is_liked）を含みます",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "投稿詳細取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "投稿ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "投稿情報",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.Post"
                        }
                    },
                    "400": {
                        "description": "無効な投稿ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "投稿が見つかりません",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/posts/{id}/like": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "指定された投稿にいいねを追加します（認証必須）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "投稿にいいね",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "投稿ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "いいねが追加された投稿",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.Post"
                        }
                    },
                    "400": {
                        "description": "無効な投稿ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "投稿が見つかりません",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "既にいいね済み",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/posts/{id}/unlike": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "指定された投稿のいいねを取り消します（認証必須）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "投稿のいいねを取り消す",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "投稿ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "いいねが取り消された投稿",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.Post"
                        }
                    },
                    "400": {
                        "description": "無効な投稿ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "投稿が見つかりません",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "いいねしていない投稿",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/uploads/images": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "複数の画像を一括アップロードし、保存されたURLの配列を返却します",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "uploads"
                ],
                "summary": "画像アップロード",
                "parameters": [
                    {
                        "type": "file",
                        "description": "アップロードする画像（複数可）",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.UploadImagesResponse"
                        }
                    },
                    "400": {
                        "description": "バリデーションエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "413": {
                        "description": "ファイルサイズ超過",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "全てのユーザーの一覧を取得します（総数付き）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザー一覧取得",
                "responses": {
                    "200": {
                        "description": "ユーザー一覧と総数",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.UsersResponse"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "指定されたIDのユーザー情報を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザー詳細取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ユーザー情報",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.User"
                        }
                    },
                    "400": {
                        "description": "無効なユーザーID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "ユーザーが見つかりません",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/{id}/posts": {
            "get": {
                "description": "指定されたユーザーの全ての投稿を取得します（総数付き）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザーの投稿一覧取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "投稿一覧と総数",
                        "schema": {
                            "$ref": "#/definitions/go-shisha-backend_internal_models.PostsResponse"
                        }
                    },
                    "400": {
                        "description": "無効なユーザーID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "ユーザーが見つかりません",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "go-shisha-backend_internal_models.AuthResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/go-shisha-backend_internal_models.User"
                }
            }
        },
...
        "go-shisha-backend_internal_models.CreateUserInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 12
                }
            }
        },
        "go-shisha-backend_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "go-shisha-backend_internal_models.Flavor": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "go-shisha-backend_internal_models.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "go-shisha-backend_internal_models.PostsResponse": {
            "type": "object",
            "properties": {
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-shisha-backend_internal_models.Post"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}